requirements.txt
langchain
openai
python-dotenv
flask
flask_sqlalchemy
flask_wtf
wtforms
flask_bcrypt
langflow